feat(parser): implement state-driven context for hierarchical Org parsing
- Introduced `ParserContext` and `PendingOrgSection` to manage document traversal states.
- Implemented incremental parent headline tracking via stack-based inheritance.
- Refactored `OrgSection` to support `Unvalidated`/`Validated` states with `Debug` and `Clone`.
- Added `ParserError` for stack integrity and mapping error management.
- Included extensive unit tests for context transitions and hierarchy resolution.
- Added `cargo-modules` to development environment for architecture visualization.

+++ rust_vector_engine/src/parser/org_kind.rs
@@ -15,12 +15,13 @@
 
 use std::ops::Range;
 
+#[derive(Debug, Clone)]
 pub enum OrgKind {
     Preamble {
         content: Range<usize>,
     },
     Headline {
-        level: u8,
+        level: usize,
         headline_range: Range<usize>,
         content_ranges: Vec<Range<usize>>,
     },
